refactor(#33): move CTR mode to a cipher modes dir
